refactor: delete unused code in Http expectations
