+++ Source/aweXpect/That/Http/StatusCodeResult.cs
@@ -133,7 +133,7 @@ public class StatusCodeResult(
 			}
 
 			string formattedResponse =
-				await ThatHttpResponseMessage.HttpResponseMessageFormatter.Format(actual, "  ", cancellationToken);
+				await HttpResponseMessageFormatter.Format(actual, "  ", cancellationToken);
 			return new ConstraintResult.Failure<HttpResponseMessage?>(actual, ToString(),
 				$"{it} had status code {Formatter.Format(value)}").WithContext("HTTP-Request", formattedResponse);
 		}
